tests/fixtures/python_repo/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "python-repo-fixture"
version = "0.1.0"
description = "Eval fixture — small FastAPI service skeleton for code-context benchmarks"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "sqlalchemy>=2.0",
    "passlib[bcrypt]>=1.7",
    "python-jose[cryptography]>=3.3",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = ["pytest>=8", "pytest-asyncio>=0.23", "httpx>=0.27"]
